+++ src/Zmsentities/Collection/ProcessList.php
@@ -115,40 +115,6 @@ class ProcessList extends Base
         return $list;
     }
 
-    public function toProcessListByStatusAndTime()
-    {
-        $list = $this->getWithHoursByDay();
-        $oldList = clone $this->sortByArrivalTime();
-        foreach ($oldList as $process) {
-            foreach ($list as $hour => $timeList) {
-                $timeList = new self();
-                $appointment = $process->getFirstAppointment();
-                if ($hour == $appointment->toDateTime()->format('H')) {
-                    if (!isset($list[$hour][intval($appointment['date'])])) {
-                        $list[$hour][intval($appointment['date'])] = $timeList;
-                    }
-                    $list[$hour][intval($appointment['date'])]->addEntity(clone $process);
-                    ksort($list[$hour]);
-                }
-            }
-        }
-        return $list;
-    }
-
-    public function getWithHoursByDay()
-    {
-        $list = array();
-        $start = 7;
-        $endTime = 18;
-        $hour = $start;
-        while ($hour <= $endTime) {
-            $list[$hour] = array();
-            $hour++;
-        }
-        ksort($list);
-        return $list;
-    }
-
     public function getScopeList()
     {
         $list = new ScopeList();